+++ app-server/subprojects/bounded_context/user/application/src/main/kotlin/club/staircrusher/user/application/port/in/use_case/LoginWithKakaoUseCase.kt
@@ -8,7 +8,7 @@ import club.staircrusher.user.application.port.`in`.InitialNicknameGenerator
 import club.staircrusher.user.application.port.`in`.UserApplicationService
 import club.staircrusher.user.application.port.out.persistence.UserAuthInfoRepository
 import club.staircrusher.user.application.port.out.persistence.UserRepository
-import club.staircrusher.user.application.port.out.web.KakaoLoginService
+import club.staircrusher.user.application.port.out.web.login.kakao.KakaoLoginService
 import club.staircrusher.user.domain.model.UserAuthInfo
 import club.staircrusher.user.domain.model.UserAuthProviderType
 import club.staircrusher.user.domain.service.UserAuthService
