+++ app-server/subprojects/spring_web/src/integrationTest/kotlin/club/staircrusher/spring_web/mock/MockKakaoLoginService.kt
@@ -2,8 +2,8 @@ package club.staircrusher.spring_web.mock
 
 import club.staircrusher.stdlib.clock.SccClock
 import club.staircrusher.stdlib.di.annotation.Component
-import club.staircrusher.user.application.port.out.web.KakaoIdToken
-import club.staircrusher.user.application.port.out.web.KakaoLoginService
+import club.staircrusher.user.application.port.out.web.login.kakao.KakaoIdToken
+import club.staircrusher.user.application.port.out.web.login.kakao.KakaoLoginService
 import org.springframework.context.annotation.Primary
 
 @Primary
