+++ app-server/subprojects/testing/spring_it/src/main/kotlin/club/staircrusher/testing/spring_it/mock/MockKakaoLoginService.kt
@@ -1,8 +1,8 @@
 package club.staircrusher.testing.spring_it.mock
 
 import club.staircrusher.stdlib.clock.SccClock
-import club.staircrusher.user.application.port.out.web.KakaoIdToken
-import club.staircrusher.user.application.port.out.web.KakaoLoginService
+import club.staircrusher.user.application.port.out.web.login.kakao.KakaoIdToken
+import club.staircrusher.user.application.port.out.web.login.kakao.KakaoLoginService
 
 class MockKakaoLoginService : KakaoLoginService {
     override fun parseIdToken(idToken: String): KakaoIdToken {
